requirements.txt
annotated-types==0.7.0
asgiref==3.8.1
blis==0.7.11
catalogue==2.0.10
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.7
cloudpathlib==0.16.0
confection==0.1.5
crispy-bootstrap5==2024.2
cymem==2.0.8
Django==4.2.3
django-crispy-forms==2.1
djangorestframework==3.15.1
idna==3.7
Jinja2==3.1.4
langcodes==3.4.0
language_data==1.2.0
marisa-trie==1.1.1
MarkupSafe==2.1.5
murmurhash==1.0.10
numpy==1.26.4
packaging==24.0
preshed==3.0.9
pydantic==2.7.2
pydantic_core==2.18.3
requests==2.32.3
smart-open==6.4.0
spacy==3.7.4
spacy-legacy==3.0.12
spacy-loggers==1.0.5
sqlparse==0.5.0
srsly==2.4.8
thinc==8.2.3
tqdm==4.66.4
typer==0.9.4
typing_extensions==4.12.1
urllib3==2.2.1
wasabi==1.1.3
weasel==0.3.4
